+++ R/app_server.R
@@ -557,8 +557,10 @@ create_app_server <- function(app_version) {
     input = input,
     output = output,
     session = session,
+    dataset_fn = dataset,
     selected_names_fn = selected_names,
     variable_table_fn = regression_variable_table,
+    category_table_fn = category_label_values,
     labels_fn = var_label_overrides,
     mark_settings_dirty = mark_settings_dirty
   )
@@ -706,6 +708,15 @@ create_app_server <- function(app_version) {
   )
 
   register_hierarchical_results_output(
+    input,
+    output,
+    analyses_fn = analyses,
+    variable_table_fn = regression_variable_table,
+    labels_fn = var_label_overrides,
+    category_table_fn = category_label_values
+  )
+
+  register_hierarchical_save_handlers(
     input,
     output,
     analyses_fn = analyses,